[DEV] Disable IMU heating for now
The IMU heating is made by enabling/disabling an heater resistor. If
APM crashes or quits for whatever reason while the IMU heating is
enabled, no counter-measure is active to disable it yet ... which
can result in physical damages
While no such counter measure is implemented, it's safer to just
disable IMU heating for now.

Change-Id: Ic0ed61ae1952aa18de1feb273f633a4cddeba68b
Reviewed-on: https://gerrit.parrot.biz/25206
Reviewed-by: Julien Béraud <julien.beraud@parrot.com>